pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pypsst"
version = "0.0.1"
description = "A module for encryption using ECC and RSA and saving contacts."
authors = [{name = "Nynra", email="nynradev@pm.me"}]
readme = "README.md"
license = {text = "BSD 3-Clause License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "pycryptodome", "typing",
]

[tool.setuptools.packages.find]
where = ["src"]
